+++ NEWS.md
@@ -8,6 +8,7 @@
 * Fix ability to upload meta data with objects (#108)
 * If file has no filesize will message without an error (#123)
 * If project has no buckets will give an empty data frame and not an error (#100)
+* Fix `gcs_save_all()` to also accept bucket level ACL (#129)
 
 # googleCloudStorageR 0.5.1